agents/framework_prototype/requirements.txt
# Core Framework Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
streamlit==1.28.0  # Dashboard UI
google-genai==1.8.0  # Gemini API support

# LLM Integration
groq==0.4.1
openai==1.3.0  # Optional: for OpenAI models
anthropic==0.7.0  # Optional: for Claude models

# Web Search & Data Fetching
requests==2.31.0
beautifulsoup4==4.12.2
duckduckgo-search==4.1.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
pyyaml==6.0.1
jsonschema==4.20.0

# State Management & Storage
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Async & Concurrency
asyncio==3.4.3
aiohttp==3.9.1
aiofiles==23.2.1

# Logging & Monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Task Queue & Background Jobs
celery==5.3.4  # Optional: for distributed tasks
redis==5.0.1  # Optional: for Celery broker

# Validation & Error Handling
tenacity==8.2.3  # Retry logic
validators==0.22.0

# CLI
click==8.1.7
rich==13.7.0  # Beautiful terminal output
typer==0.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2  # For testing FastAPI

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6  # For file uploads
jinja2==3.1.2

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Optional: Apache Integration (for future phases)
# kafka-python==2.0.2
# apache-airflow==2.7.3
# apache-airflow-providers-celery==3.4.1

# Optional: Advanced Features
# langchain==0.0.350  # For inspiration/comparison
# chromadb==0.4.18  # Vector DB for memory
# faiss-cpu==1.7.4  # Alternative vector store